refactor: ⚡️ プレイヤーのセットをフックを介して行うように変更
